third_party/blink/web_tests/external/wpt/websockets/Create-Secure-extensions-empty.any-expected.txt

This is a testharness.js-based test.
FAIL Create Secure WebSocket - wsocket.extensions should be set to '' after connection is established - Connection should be opened assert_equals: extensions should be empty expected "" but got "permessage-deflate"
PASS Create Secure WebSocket - wsocket.extensions should be set to '' after connection is established - Connection should be closed
Harness: the test ran to completion.
